+++ payroll-case-study/core/src/main/groovy/io/github/thomashan/payroll/eventhandler/add/EmployeeCreatedHandler.groovy
@@ -1,4 +1,4 @@
-package io.github.thomashan.payroll.eventhandler
+package io.github.thomashan.payroll.eventhandler.add
 
 import io.github.thomashan.event.EventHandler
 import io.github.thomashan.payroll.Employee
